Created the player class, added equip method, and attack method
Created a simple player class with a name, health, level, and equpedItems. I added an equip method for the player to holding the item before attacking. Also created the attack method where is checks if the item has any damage points and for now prints out a message, but when we create the bosses then the attack method will damage the enemy's health.